feat(#4): stations permissions

+++ app/controllers/stationController.js
@@ -1,20 +1,27 @@
 const Station = require ('../models/Station');
+const User = require ('../models/User');
 
-const createStation = (req, res) => {
-    const station = new Station({
-        name: req.body.name,
-        openHour: req.body.openHour,
-        closeHour: req.body.closeHour
-    });
+const createStation = async (req, res) => {
+    const currentUser = await User.findOne({_id: req.userId});
 
-    station.save()
-        .then(result => {
-         res.status(201).json(result);
-       })
-        .catch (err => {
-            console.log(err);
-            res.status(500).json({ error: 'Internal Server Error' });
-       });
+    if (currentUser.role == "ADMIN") {
+        const station = new Station({
+            name: req.body.name,
+            openHour: req.body.openHour,
+            closeHour: req.body.closeHour
+        });
+
+        station.save()
+            .then(result => {
+            res.status(201).json(result);
+        })
+            .catch (err => {
+                console.log(err);
+                res.status(500).json({ error: 'Internal Server Error' });
+        });
+    } else {
+        res.status(403).json({ error: 'You do not have the suffisant privilieges to perform this action'})
+    }
 };
 
 const getAllStations = (req, res) => {
@@ -43,33 +50,45 @@ const getSingleStation = (req, res) => {
         });
 };
 
-const updateStation = (req, res) => {
-    const id = req.params.id;
+const updateStation = async (req, res) => {
+    const currentUser = await User.findOne({_id: req.userId});
 
-    Station.findByIdAndUpdate(id , req.body)
-        .then(result => {
-            if (result) {
-                res.status(200).send('Updated station successfully'); // 200 OK
-            } else {
-                res.status(404).json({ error: 'Station not found' }); // 404 Not Found
-            }
-        })
-        .catch(err => {
-            console.error(err);
-            res.status(500).json({ error: 'Internal Server Error' }); // 500 Internal Server Error
-        });
+    if (currentUser.role == "ADMIN") {
+        const id = req.params.id;
+
+        Station.findByIdAndUpdate(id , req.body)
+            .then(result => {
+                if (result) {
+                    res.status(200).send('Updated station successfully'); // 200 OK
+                } else {
+                    res.status(404).json({ error: 'Station not found' }); // 404 Not Found
+                }
+            })
+            .catch(err => {
+                console.error(err);
+                res.status(500).json({ error: 'Internal Server Error' }); // 500 Internal Server Error
+            });
+    } else {
+        res.status(403).json({ error: 'You do not have the suffisant privilieges to perform this action'})
+    }
 };
 
-const deleteStation = (req, res) => {
-    const id = req.params.id;
+const deleteStation = async (req, res) => {
+    const currentUser = await User.findOne({_id: req.userId});
 
-    Station.findByIdAndDelete(id)
-        .then(result => {
-            res.send('Station deleted successful');
-        })
-        .catch(err => {
-            console.log(err);
-        });
+    if (currentUser.role == "ADMIN") {
+        const id = req.params.id;
+
+        Station.findByIdAndDelete(id)
+            .then(result => {
+                res.send('Station deleted successful');
+            })
+            .catch(err => {
+                console.log(err);
+            });
+    } else {
+        res.status(403).json({ error: 'You do not have the suffisant privilieges to perform this action'})
+    }
 };
 
 module.exports = { createStation , getAllStations, getSingleStation ,updateStation, deleteStation };
